requirements.txt
python-dotenv==0.21.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
